# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/current": {
            "get": {
                "description": "Downloads weather data for a location for the past N days (default 1 day)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "weather"
                ],
                "summary": "Download and update recent weather data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "City name",
                        "name": "place",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Latitude of the location",
                        "name": "latitude",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Longitude of the location",
                        "name": "longitude",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of days in the past to fetch data for (default is 1)",
                        "name": "count",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Measurement"
                            }
                        }
                    }
                }
            }
        },
        "/generate": {
            "get": {
                "description": "Generates and stores N random weather records (default is 1)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "weather"
                ],
                "summary": "Generate random weather data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of random records to generate (default is 1)",
                        "name": "count",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "null response",
                        "schema": {}
                    }
                }
            }
        },
        "/rain-intensity": {
            "get": {
                "description": "Retrieves cities that experienced rain of the specified intensity within a time interval",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "weather"
                ],
                "summary": "Get cities by rain intensity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date in YYYY-MM-DD format",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date in YYYY-MM-DD format",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Rain intensity threshold (e.g., mm/hour)",
                        "name": "intensity",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.City"
                            }
                        }
                    }
                }
            }
        },
        "/retrieve": {
            "get": {
                "description": "Retrieves weather measurements for a given city or coordinates within a specified time interval",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "weather"
                ],
                "summary": "Retrieve weather measurements",
                "parameters": [
                    {
                        "type": "string",
                        "description": "City name",
                        "name": "place",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Latitude of the location",
                        "name": "latitude",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Longitude of the location",
                        "name": "longitude",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start date in YYYY-MM-DD format",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date in YYYY-MM-DD format",
                        "name": "to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Measurement"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.City": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "find_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.Measurement": {
            "type": "object",
            "properties": {
                "city_id": {
                    "type": "integer"
                },
                "ground_level": {
                    "type": "number"
                },
                "humidity": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "max_temperature": {
                    "type": "number"
                },
                "min_temperature": {
                    "type": "number"
                },
                "pressure": {
                    "type": "number"
                },
                "rain_intensity": {
                    "type": "number"
                },
                "sea_level": {
                    "type": "number"
                },
                "temperature": {
                    "type": "number"
                },
                "timestamp": {
                    "type": "string"
                },
                "weather": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.MeasurementWeather"
                    }
                },
                "wind_degrees": {
                    "type": "number"
                },
                "wind_speed": {
                    "type": "number"
                }
            }
        },
        "model.MeasurementWeather": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "main": {
                    "type": "string"
                }
            }
        }
    }
}